+++ R/1.R
@@ -52,6 +52,7 @@ emulate <- function(object,...) UseMethod("emulate")
 AICc <- function(object,...) UseMethod("AICc")
 speed <- function(object,...) UseMethod("speed")
 speeds <- function(object,...) UseMethod("speeds")
+mag <- function(x,...) UseMethod("mag")
 modes <- function(object,...) UseMethod("modes")
 
 # internal S3 generic function